DebugOverlay improvements
Use rebuild 'DestroyOverlay()'
Use 'size_t' type for MemAlloc_Internal alloc size.

+++ r5dev/tier0/tslist.cpp
@@ -1,7 +1,7 @@
 #include "core/stdafx.h"
 #include "tier0/tslist.h"
 
-void* MemAlloc_Internal(void* pPool, int64_t nSize)
+void* MemAlloc_Internal(void* pPool, size_t nSize)
 {
 	return v_MemAlloc_Internal(pPool, nSize);
 }
